do JSON data in Javascript
WebCola wants the 'links' JSON hash to refer to commits by
their index in the 'nodes' hash.  There are two issues with this:

1. This is far less useful for any consumers of the JSON generated by
   the --json option in CLI mode.

2. More importantly, when implementing a graph visualization which can
   incremently add nodes, the index of a commit in the 'nodes' hash
   depends on what was already there client-side in the browser.

So instead we refer to commits by SHA server-side (in the Python) and
then convert those to indices within the 'nodes' hash client-side via
Javascript.  This effectively means moving the self._nodes hash from
Python to Javascript.