change coupling of detector with listeners
Get rid of the special case where a listener can be added to a detector
on construction.  This allows us to construct the detector before any
listeners, and then invoking add_listener() can take care of creating
bi-directional links, which allows any listener to ask the detector for
the repository being examined (and hence access the full pygit2 API for
that repository).