ditch externs hack
It was based on flawed thinking.  If the variable being exported gets
reassigned a fresh object, having it wrapped in an externs object
doesn't help.  And if you always have to repoint the export at the newly
initialised object, having an extra externs wrapper doesn't win
anything.  And it loses by requiring an extra level of indirection.