notify of new dependency even if target already done
When a new dependency is discovered, its target may have already been
placed on the TODO list via a different source (and maybe even analysed
for dependencies), e.g. if find_dependencies() was previously called on
it, or if it was previously found via recursion.  In this case we still
need to notify listeners of the dependency, even though we won't add it
to the TODO list.